Move to 1.9 int

+++ src/Orchard.Web/Modules/Orchard.Indexing/Services/IndexingTaskManager.cs
@@ -26,11 +26,17 @@ namespace Orchard.Indexing.Services {
         public ILogger Logger { get; set; }
 
         private void CreateTask(ContentItem contentItem, int action) {
-            if ( contentItem == null ) {
+            if (contentItem == null) {
                 throw new ArgumentNullException("contentItem");
             }
 
-            foreach (var task in _repository.Table.Where(task => task.ContentItemRecord == contentItem.Record)) {
+            if (contentItem.Record == null) {
+                // ignore that case, when Update is called on a content item which has not be "created" yet
+
+                return;
+            }
+
+            foreach (var task in _repository.Fetch(task => task.ContentItemRecord == contentItem.Record)) {
                 _repository.Delete(task);
             }
